+++ inst/include/rxode2dataErr.h
@@ -25,6 +25,7 @@
 #define rxErrModeledFss2n2    2097152
 #define rxErrModeledFss2n3    4194304
 #define rxErrRate02           8388608
+#define rxErrIndLinConverge   16777216
 
 #define rxErrNaTimeLag   1
 #define rxErrNaTimeRate  2